+++ Services/IAlbaranService.cs
@@ -6,8 +6,8 @@ namespace MauiApp1.Api.Services
 {
     public interface IAlbaranService
     {
-        Task<Albaran?> GetAlbaranOrigenAsync(string serie, int numDoc);
-        Task<Albaran?> GetAlbaranDestinoAsync(string serie, int numDoc);
+        Task<Albaran?> GetAlbaranOrigenAsync(string serie, decimal numDoc);
+        Task<Albaran?> GetAlbaranDestinoAsync(string serie, decimal numDoc);
         Task<List<Albaran>> GetAllAlbaranesAsync();
         Task<bool> AddAlbaranAsync(Albaran albaran);
         Task<bool> UpdateAlbaranOrigenAsync(int albaranId, Albaran updatedData);
